+++ src/soil.cpp
@@ -396,7 +396,9 @@ NumericVector temperatureChange(NumericVector dVec, NumericVector Temp,
 }
 
 // [[Rcpp::export("soil")]]
-List soil(DataFrame SoilParams, String VG_PTF = "Toth", NumericVector W = NumericVector::create(1.0), double SWE = 0.0) {
+List soil(DataFrame SoilParams, String VG_PTF = "Toth", 
+          NumericVector W = NumericVector::create(1.0), 
+          double SWE = 0.0) {
   double SoilDepth = 0.0;
   NumericVector dVec = clone(as<NumericVector>(SoilParams["widths"]));
   int nlayers = dVec.size();
@@ -446,9 +448,11 @@ List soil(DataFrame SoilParams, String VG_PTF = "Toth", NumericVector W = Numeri
     VG_theta_sat[l] = vgl[3];
     // Stolf, R., Thurler, A., Oliveira, O., Bacchi, S., Reichardt, K., 2011. Method to estimate soil macroporosity and microporosity based on sand content and bulk density. Rev. Bras. Ciencias do Solo 35, 447–459.
     macro[l] = std::max(0.0,0.693 - 0.465*bd[l] + 0.212*(sand[l]/100.0));
-    Ksat[l] = saturatedConductivitySaxton(sand[l], clay[l], om[l]);
+    Ksat[l] = saturatedConductivitySaxton(clay[l], sand[l], om[l]);
     SoilDepth +=dVec[l];
   }
+  // Saturated vertical hydraulic conductivity (mm/day) 
+  double Kperc = 0.05*saturatedConductivitySaxton(clay[nlayers-1], sand[nlayers-1], om[nlayers-1], false);
   double Ksoil = 0.05;
   double Gsoil = 0.5; //TO DO, implement pedotransfer functions for Gsoil
   List l = List::create(_["SoilDepth"] = SoilDepth,
@@ -458,10 +462,10 @@ List soil(DataFrame SoilParams, String VG_PTF = "Toth", NumericVector W = Numeri
                       _["Ksoil"] = Ksoil, _["Gsoil"] = Gsoil,
                       _["dVec"] = dVec,
                       _["sand"] = sand, _["clay"] = clay, _["om"] = om,
-                      _["usda_Type"] = usda_Type,
                       _["VG_alpha"] = VG_alpha,_["VG_n"] = VG_n, 
                       _["VG_theta_res"] = VG_theta_res,_["VG_theta_sat"] = VG_theta_sat,
                       _["Ksat"] = Ksat,
+                      _["Kperc"] = Kperc,
                       _["macro"] = macro,
                       _["bd"] = bd,
                       _["rfc"] = rfc);